+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -2801,6 +2801,52 @@ class SentryClientTest {
     }
 
     @Test
+    fun `does not captureReplay for cached events`() {
+        var called = false
+        fixture.sentryOptions.setReplayController(object : ReplayController by NoOpReplayController.getInstance() {
+            override fun captureReplay(isTerminating: Boolean?) {
+                called = true
+            }
+        })
+        val sut = fixture.getSut()
+
+        sut.captureEvent(
+            SentryEvent().apply {
+                exceptions = listOf(
+                    SentryException().apply {
+                        mechanism = Mechanism().apply { isHandled = false }
+                    }
+                )
+            },
+            HintUtils.createWithTypeCheckHint(CachedHint())
+        )
+        assertFalse(called)
+    }
+
+    @Test
+    fun `captures replay for cached events with apply scope`() {
+        var called = false
+        fixture.sentryOptions.setReplayController(object : ReplayController by NoOpReplayController.getInstance() {
+            override fun captureReplay(isTerminating: Boolean?) {
+                called = true
+            }
+        })
+        val sut = fixture.getSut()
+
+        sut.captureEvent(
+            SentryEvent().apply {
+                exceptions = listOf(
+                    SentryException().apply {
+                        mechanism = Mechanism().apply { isHandled = false }
+                    }
+                )
+            },
+            HintUtils.createWithTypeCheckHint(CachedWithApplyScopeHint())
+        )
+        assertTrue(called)
+    }
+
+    @Test
     fun `when beforeSendReplay is set, callback is invoked`() {
         var invoked = false
         fixture.sentryOptions.setBeforeSendReplay { replay: SentryReplayEvent, _: Hint -> invoked = true; replay }
@@ -3123,6 +3169,10 @@ class SentryClientTest {
     private class BackfillableHint : Backfillable {
         override fun shouldEnrich(): Boolean = false
     }
+
+    private class CachedHint : Cached
+
+    private class CachedWithApplyScopeHint : Cached, ApplyScopeData
 }
 
 class DropEverythingEventProcessor : EventProcessor {